UPSTREAM: 48394: GuaranteedUpdate must write if stored data is not canonical
An optimization added to the GuaranteedUpdate loop changed the
comparison of the current objects serialization against the stored data,
instead comparing to the in memory object, which defeated the mechanism
we use to migrate stored data.

This commit preserves that optimization but correctly verifies the in
memory serialization against the on disk serialization by fetching the
latest serialized data. Since most updates are not no-ops, this should
not regress the performance of the normal path.

:100644 100644 210b67084c... 708018cd49... M	pkg/storage/etcd3/store.go
:100644 100644 cc1d76c778... ed9c3bf3c8... M	pkg/storage/etcd3/store_test.go