UPSTREAM: 63716: Add InstallPathHandler which allows for more then one path to be associated with health checking.
Currently it is only possible to have one group of checks which must all pass for the handler to report success.
Allowing multiple paths for these checks allows use of the same machinery for other kinds of checks, i.e. readiness.

This upstream change allows for the differentiation of health and readiness checks

Origin-commit: 417e073f09096f54974022b5381606a38bfa68b1

Kubernetes-commit: bf1ea267be633274d9be816b88537e55156a839a